pyproject.toml
[project]
name = "chatbot-multimodal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-cognitiveservices-speech>=1.43.0",
    "azureml>=0.2.7",
    "black>=25.1.0",
    "chainlit>=2.5.5",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "openai>=1.77.0",
    "opencensus-ext-azure>=1.1.14",
    "pdfminer-six>=20250506",
    "pillow>=11.2.1",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "pymupdf>=1.25.5",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.8",
    "unstructured>=0.17.2",
    "uvicorn>=0.34.2",
]

[tool.black]
line-length = 88
target-version = ["py310"]
skip-string-normalization = false  # Use double quotes by default

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]  # E: pycodestyle, F: pyflakes, I: isort
ignore = []

[tool.ruff.format]
quote-style = "double"



